signals: Add support for starting and stopping cefsrc from js
See the sample in the html/ directory for an example usage. This should
be useful for those webpages that have non-trivial load times (webgl
apps, etc) - allowing the page to signal from javascript to the gst
pipeline that the page is ready for display. We can also send an EOS
signal to be turned into an event on cefsrc.

Flag for enabling this feature is listen-for-js-signals, and then on the
javascript side you can send signals (see html sample for detail):

window.gstSendMsg({ request: "ready" | "eos", ...})